# src/main/resources/openapi.yaml
openapi: '3.0.1'

info:
  title: Vana Logistiek App
  description: Een ticket-systeem voor de supporting teams op Castlefest
  version: 1.0.0
  license:
    name: GNU General Public License 3.0

servers:
  - description: Quarkus Dev Mode
    url: http://localhost:8000/


security:
  - jwt

paths:
  /auth/login:
    summary: Creates a new session
    post:
      operationId: PostLogin
      summary: Starts a new session, returns a JSON Web Token
      tags:
        - Authentication
      security: [ ]
      requestBody:
        $ref: '#/components/requestBodies/PostLoginRequest'
      responses:
        200:
          description: Succesful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLoginResponse"
        400:
          description: Credentials were incorrect
        419:
          description: Rate Limited, try again after a little while

  /auth/me:
    summary: Returns information about the session
    get:
      operationId: GetMe
      summary: Returns user information
      tags:
        - Authentication
      responses:
        200:
          description: OK
          content:
            application/json:
              example:
                name: Kim Wilde
              schema:
                $ref: "#/components/schemas/GetMeResponse"
        401:
          $ref: "#/components/responses/401"

components:
  requestBodies:
    PostLoginRequest:
      description: A request to login
      required: true
      content:
        application/json:
          example:
            username: john.doe
            password: hunter2
          schema:
            title: PostLoginRequest
            type: object
            required:
              - username
              - password
            additionalProperties: false
            properties:
              username:
                type: string
                nullable: false
              password:
                type: string
                nullable: false

  responses:
    401:
      description: Not logged in
      content:
        application/json:
          example:
            code: '401.001'
            message: 'You are not logged in'
          schema:
            $ref: "#/components/schemas/error"

  schemas:
    error:
      type: object
      required:
        - code
        - message
      additionalProperties: true
      properties:
        code:
          type: string
          description: The error code used to distinguish this error
          example: '400.001'
        message:
          type: string
          description: A detailed description
          example: 'The field [field] is required.'
    PostLoginResponse:
      type: object
      required:
        - name
        - jwt
        - expiration
      additionalProperties: false
      properties:
        name:
          type: string
          nullable: true
        jwt:
          type: string
          format: jwt
        expiration:
          type: string
          format: datetime
    GetMeResponse:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email

  securitySchemes:
    jwt:
      description: Quarkus session
      type: http
      scheme: bearer
      bearerFormat: JSON Web Token